Moving type guard isGetIdentityData method after its declaration
